backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors==0.1
django-environ==0.11.2
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
